Refactor Terraform command builders into separate files
